Validate Ent input fields are only one ent field
The `#[ent(...)]` attribute syntax is convenient and readable, but has
some implicit requirements that are not expressed in the API syntax.
The macro already enforced the requirement that known fields such as
`id` can only be declared once per struct, but it didn't enforce that
a single input field can only be one ent field.

This commit adds that validation, and amends the existing validation to
ensure that no spurious "missing field" errors are emitted due to the
order of branches in field analysis.

Fixes #30